AI/requirements.txt
agno
pgvector
sqlalchemy
langchain-text-splitters
pypdf
fastapi
uvicorn
python-dotenv
pydantic
requests
transformers
torch
pymupdf
streamlit
openai
mistralai
groq
einops
langchain_community
serpapi
crawl4ai
asyncio
langchain
google-search-results
playwright
playwright[chromium]
aiohttp
psycopg2-binary
google-genai
SpeechRecognition
gTTS
pydub
